Robot code
This is more or less the final robot code. As it currently stands, we
have no image processing, the joystick constants have been adjusted, and
the IDs may need to be changed. The code for the camera will need to be
uncommented if we actually put it on the final robot. Also, the
autonomous has been mostly tested but may need some final adjustments