Robot Code 2/22
Autonomous reverted to use the sensors. We will be selecting auto before
by discussing with teams what we will do. The sensor auto code's logic
needs to be tested, but until then it would be safest to rely on the
code that simply drives up and simply goes over. Spy code still needs to
be tested